# back-end/swagger.yaml
openapi: 3.0.3
info:
  title: APIs tisco - OpenAPI 3.0
  description: APIs for tisco
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
  - url: http://159.138.242.188:3000/api/v1
...
components:
  securitySchemes:
    beareAuth:
        type: http
        scheme: bearer
        name: Authorization
        in: header
        bearerFormat: JWT
        description: >-
          Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345"
...
paths:
...
  /require/sent-otp:
    post:
      tags:
        - require
      security:
        - beareAuth: []
      summary: Send OTP
      operationId: sendOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [OTP]
                input:
                  type: string
              required:
                - action
                - input
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  secret:
                    type: string
                  type:
                    type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /require/verify-otp:
    post:
      tags:
        - require
      security:
        - beareAuth: []
      summary: Verify OTP
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                otp:
                  type: string
                secret:
                  type: string
              required:
                - otp
                - secret
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  isValid:
                    type: boolean
                  response_code:
                    type: string
                  message:
                    type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string   
...
  /game/save-answer:
    post:
      tags:
        - game-board
      security:
        - beareAuth: []
      summary: Save game board answer
      operationId: createHistory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gmBoardId: 
                  type: number
                gmGameLevel:
                  type: number
                gmScore:  
                  type: number
                gmHisSystem:
                  type: object
                gmHisRecord:
                  type: array
                  items:
                    type: object
                    properties:
                      question:
                        type: string
                      answers:
                        type: array
                        items:
                          type: object
                          properties:
                            answer:
                              type: string
                            ansScore:
                              type: integer
                    required:
                    - question
                    - answers
              required:
                - gmBoardId
                - gmGameLevel
                - gmScore
                - gmHisSystem
                - gmHisRecord
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string